refactor: move load_example_data test to test_utils.py
